pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ser2tcp"
dynamic = ["version"]
description = "Serial port proxy to TCP or TELNET"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    { name = "Pavel Revak", email = "pavel.revak@gmail.com" }
]
keywords = ["serial", "tcp", "telnet", "proxy"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Embedded Systems",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pyserial>=3.0",
    "uhttp-server>=2.3.2",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/cortexm/ser2tcp"

[project.scripts]
ser2tcp = "ser2tcp.main:main"

[tool.setuptools.packages.find]
include = ["ser2tcp*"]

[tool.setuptools.package-data]
ser2tcp = ["html/**"]

[tool.setuptools_scm]
write_to = "ser2tcp/_version.py"
fallback_version = "0.0.0"
